requirements.txt
# Base
requests==2.31.0
pyyaml==6.0.1
tqdm==4.66.1
python-dotenv==1.0.0
pydantic>=2.0.0

# API Clients
openai>=1.3.0
azure-identity==1.15.0
azure-ai-textanalytics==5.3.0

# Data Processing
beautifulsoup4==4.12.2
lxml==4.9.3
nltk==3.8.1
spacy==3.7.2
pandas==2.1.1

# Testing
pytest==7.4.3

# NLP
transformers>=4.30.0
sentencepiece>=0.1.99
torch>=2.2.0